+++ package.json
@@ -13,6 +13,7 @@
 		"db:push": "drizzle-kit migrate",
 		"db:generate": "drizzle-kit generate",
 		"db:migrate": "drizzle-kit migrate",
+		"email:dev": "email dev -p 5005",
 		"storybook": "storybook dev -p 6006",
 		"build-storybook": "storybook build"
 	},
@@ -45,6 +46,7 @@
 	},
 	"devDependencies": {
 		"@biomejs/biome": "2.3.11",
+		"@react-email/preview-server": "5.2.10",
 		"@storybook/nextjs-vite": "^10.2.8",
 		"@tailwindcss/postcss": "^4",
 		"@types/node": "^20",
@@ -55,6 +57,7 @@
 		"dotenv-cli": "^11.0.0",
 		"drizzle-kit": "^0.31.8",
 		"jsdom": "^27.4.0",
+		"react-email": "5.2.10",
 		"storybook": "^10.2.8",
 		"tailwindcss": "^4",
 		"tsx": "^4.21.0",
